Wybór następnego modułu - w trakcie

+++ EduBot/EduApi/EduApi/DTO/ModuleDTO.cs
@@ -6,7 +6,7 @@ namespace EduApi.DTO {
     public class ModuleDTO {
 
         public int id { get; set; }
-        public short? group_id { get; set; }
+        public int? group_id { get; set; }
         public int group_position { get; set; }
         public string difficulty { get; set; }
         public string title { get; set; }
